+++ src/dchshx.f90
@@ -88,9 +88,10 @@
 !>
 !> \ingroup choldecomp
 subroutine dchshx(n,R,ldr,i,j,w)
+  use iso_fortran_env
     integer, intent(in) :: n, ldr, i, j
-    double precision, intent(inout) :: R(ldr,*)
-    double precision, intent(out) :: w(*)
+    real(real64), intent(inout) :: R(ldr,*)
+    real(real64), intent(out) :: w(*)
     external xerbla,dcopy,dqrtv1,dqrqh,dqhqr
     integer info,l
     ! quick return if possible.
