+++ src/dqrinr.f90
@@ -105,11 +105,12 @@
 !>
 !> \ingroup qrdecomp
 subroutine dqrinr(m,n,Q,ldq,R,ldr,j,x,w)
+  use iso_fortran_env
     integer, intent(in) :: m, n, j, ldq, ldr
-    double precision, intent(inout) :: Q(ldq,*)
-    double precision, intent(inout) :: R(ldr,*)
-    double precision, intent(inout) :: x(*)
-    double precision, intent(out) :: w(*)
+    real(real64), intent(inout) :: Q(ldq,*)
+    real(real64), intent(inout) :: R(ldr,*)
+    real(real64), intent(inout) :: x(*)
+    real(real64), intent(out) :: w(*)
     external xerbla,dcopy,dqhqr,dqrot
     integer info,i,k
     ! check arguments
