+++ src/dqrshc.f90
@@ -114,10 +114,11 @@
 !>
 !> \ingroup qrdecomp
 subroutine dqrshc(m,n,k,Q,ldq,R,ldr,i,j,w)
+  use iso_fortran_env
     integer, intent(in) :: m, n, k, ldq, ldr, i, j
-    double precision, intent(inout) :: Q(ldq,*)
-    double precision, intent(inout) :: R(ldr,*)
-    double precision, intent(out) :: w(*)
+    real(real64), intent(inout) :: Q(ldq,*)
+    real(real64), intent(inout) :: R(ldr,*)
+    real(real64), intent(out) :: w(*)
     external dqrot, xerbla,dcopy,dqrtv1,dqrqh,dqhqr
     integer info,jj,kk,l
     ! quick return if possible.
